クールタイム、硬直の更新、体力の児童回復にTime.deltaTImeを使うようにした

+++ Assets/Scripts/Characters/Base/CharData.cs
@@ -8,17 +8,17 @@ public class CharData : ScriptableObject
     public int max_burst = 100;
 
     [Header("©“®‰ñ•œ")]
-    public int regen_burst_cooltime = 120;
-    public int restart_regen_burst_value = 60;
+    public float regen_burst_cooltime = 120;
+    public float restart_regen_burst_value = 60;
 
     [Header("Skill 1")]
-    public int skill_1_rigid = 0;
-    public int skill_1_cooltime = 0;
+    public float skill_1_rigid = 0;
+    public float skill_1_cooltime = 0;
 
     [Header("Skill 2")]
-    public int skill_2_rigid = 0;
-    public int skill_2_cooltime = 0;
+    public float skill_2_rigid = 0;
+    public float skill_2_cooltime = 0;
 
     [Header("Skill 2")]
-    public int dash_cooltime = 0;
+    public float dash_cooltime = 0;
 }
